refactor waiting and clean up
