+++ datamapplot/interactive_tests/tests/arxiv_ml.spec.ts
@@ -1,16 +1,10 @@
 import { test, expect } from '@playwright/test';
+import { waitForDeckGL, waitForCanvas } from '../utils/canvas';
 
 test.describe('Arxiv ML Canvas Tests', () => {
   test.beforeEach(async ({ page }, testInfo) => {
     // Extend timeout for all tests running this hook by 4 minutes.
     testInfo.setTimeout(testInfo.timeout + 240_000);
-
-    // Add console listeners first
-    // page.on('console', msg => {
-    //   const type = msg.type();
-    //   const text = msg.text();
-    //   console.log(`Browser ${type}: ${text}`);
-    // });
 
     const response = await page.goto('http://localhost:8000/arxiv_ml.html', { timeout: 60_000 });
     expect(response.status()).toBe(200);
@@ -21,112 +15,9 @@ test.describe('Arxiv ML Canvas Tests', () => {
       page.waitForSelector('#progress-container', { state: 'hidden', timeout: 180_000 }),
     ]);
 
-    const waitForDeckGL = async (page) => {
-      console.log('Waiting for deck.gl...');
-
-      const canvas = page.locator('#deck-container canvas');
-      await canvas.waitFor({ state: 'visible', timeout: 180_000 });
-
-      const deckReady = await page.evaluate(() => {
-        const canvas = document.querySelector('#deck-container canvas');
-        if (!canvas) return { ready: false, error: 'No canvas' };
-
-        const gl = canvas.getContext('webgl2') || canvas.getContext('webgl');
-        if (!gl) return { ready: false, error: 'No WebGL context' };
-
-        // Check if deck is initialized without window assignment
-        const deckIsWorking = canvas.width > 0 && canvas.height > 0 && gl.drawingBufferWidth > 0;
-
-        return {
-          ready: deckIsWorking,
-          dimensions: {
-            width: canvas.width,
-            height: canvas.height
-          },
-          debug: {
-            deckExists: 'deck' in window,
-            canvasSize: `${canvas.width}x${canvas.height}`,
-            glSize: `${gl.drawingBufferWidth}x${gl.drawingBufferHeight}`
-          }
-        };
-      });
-
-      console.debug('Deck.gl debug state:', deckReady);
-      return canvas;
-    };
-
-    const canvas = await waitForDeckGL(page);
-
-  //  // Patch deck.gl to log redraw calls
-  //   const patchResult = await page.evaluate(() => {
-  //     try {
-  //       if (!window.deck?.Deck) {
-  //         return { success: false, error: 'No deck.gl' };
-  //       }
-
-  //       const proto = window.deck.Deck.prototype;
-  //       const origRedraw = proto.redraw;
-
-  //       proto.redraw = function(...args) {
-  //         console.log('Redraw called:', {
-  //           hasViewState: !!this.viewState,
-  //           viewState: this.viewState,
-  //           timestamp: Date.now(),
-  //           props: this.props
-  //         });
-  //         return origRedraw.apply(this, args);
-  //       };
-
-  //       return {
-  //         success: true,
-  //         method: 'redraw'
-  //       };
-  //     } catch (e) {
-  //       return {
-  //         success: false,
-  //         error: e.toString()
-  //       };
-  //     }
-  //   });
-  //   console.log('Patch result:', patchResult);
+    const deckReady = await waitForDeckGL(page, testInfo);
+    console.debug('Deck.gl state:', deckReady);
   });
-
-  const waitForCanvas = async (page) => {
-    console.log('Waiting for canvas...');
-
-    const canvas = page.locator('#deck-container canvas');
-    await canvas.waitFor({ state: 'visible', timeout: 180_000 });
-    await page.waitForLoadState('networkidle');
-    await page.waitForTimeout(500); // Additional wait for stability
-
-    // Evaluate inside the browser context to access canvas properties
-    const canvasInfo = await page.evaluate(() => {
-      const canvasSelector = document.querySelector('#deck-container canvas');
-      if (canvasSelector.width === 0 || canvasSelector.height === 0 || !canvasSelector) {
-          console.log('Try redrawing');
-          window.dispatchEvent(new Event('resize'));
-          window.dispatchEvent(new Event('redraw'));
-        }
-
-      if (!canvasSelector) return null;
-
-      const ctx = canvasSelector.getContext('webgl2') || canvasSelector.getContext('webgl');
-      if (!ctx) return {
-        width: canvasSelector.width,
-        height: canvasSelector.height,
-        error: 'No WebGL context available' };
-
-      return {
-        width: canvasSelector.width,
-        height: canvasSelector.height,
-        contextAttributes: !!ctx.getContextAttributes(),
-        extensions: !!ctx.getSupportedExtensions()
-      };
-    });
-
-    // console.debug(canvasInfo);
-    return canvas;
-  };
 
   const verifyInitialState = async (page) => {
     const canvas = await waitForCanvas(page);
