+++ datamapplot/interactive_tests/tests/cord_canvas.spec.ts
@@ -1,96 +1,24 @@
 import { test, expect } from '@playwright/test';
+import { waitForDeckGL, waitForCanvas } from '../utils/canvas';
+
 
 test.describe('Cord19 Canvas Tests', () => {
   test.beforeEach(async ({ page }, testInfo) => {
     // Extend timeout for all tests running this hook by 6 minutes.
     testInfo.setTimeout(testInfo.timeout + 360_000);
 
-    // Set higher maxDiffPixels
-    const isMobile = testInfo.project.name.includes('mobile');
-    test.expect.toHaveScreenshot.maxDiffPixels = isMobile ? 10_000 : 20_000;
-
     const response = await page.goto('http://localhost:8000/cord19.html', { timeout: 60_000 });
     expect(response.status()).toBe(200);
 
-    console.log('Waiting for initial load:', testInfo.project.name);
+    console.log('Waiting for initial load...', testInfo.project.name);
     await Promise.all([
       page.waitForSelector('#loading', { state: 'hidden', timeout: 180_000 }),
       page.waitForSelector('#progress-container', { state: 'hidden', timeout: 180_000 }),
     ]);
 
-    const waitForDeckGL = async (page) => {
-      console.debug('Waiting for deck.gl...');
-
-      const canvas = page.locator('#deck-container canvas');
-      await canvas.waitFor({ state: 'visible', timeout: 180_000 });
-
-      const deckReady = await page.evaluate(() => {
-        const canvas = document.querySelector('#deck-container canvas');
-        if (!canvas) return { ready: false, error: 'No canvas' };
-
-        const gl = canvas.getContext('webgl2') || canvas.getContext('webgl');
-        if (!gl) return { ready: false, error: 'No WebGL context' };
-
-        // Check if deck is initialized without window assignment
-        const deckIsWorking = canvas.width > 0 && canvas.height > 0 && gl.drawingBufferWidth > 0;
-
-        return {
-          ready: deckIsWorking,
-          dimensions: {
-            width: canvas.width,
-            height: canvas.height
-          },
-          debug: {
-            deckExists: 'deck' in window,
-            canvasSize: `${canvas.width}x${canvas.height}`,
-            glSize: `${gl.drawingBufferWidth}x${gl.drawingBufferHeight}`
-          }
-        };
-      });
-
-      console.debug('Deck.gl state:', deckReady);
-      return canvas;
-    };
-
-    const canvas = await waitForDeckGL(page);
+    const deckReady = await waitForDeckGL(page, testInfo);
+    console.debug('Deck.gl state:', deckReady);
   });
-
-  const waitForCanvas = async (page) => {
-    console.log('Waiting for canvas...');
-
-    const canvas = page.locator('#deck-container canvas');
-    await canvas.waitFor({ state: 'visible', timeout: 180_000 });
-    await page.waitForLoadState('networkidle');
-    await page.waitForTimeout(500); // Additional wait for stability
-
-    // Evaluate inside the browser context to access canvas properties
-    const canvasInfo = await page.evaluate(() => {
-      const canvasSelector = document.querySelector('#deck-container canvas');
-      if (canvasSelector.width === 0 || canvasSelector.height === 0 || !canvasSelector) {
-          console.log('Try redrawing');
-          window.dispatchEvent(new Event('resize'));
-          window.dispatchEvent(new Event('redraw'));
-        }
-
-      if (!canvasSelector) return null;
-
-      const ctx = canvasSelector.getContext('webgl2') || canvasSelector.getContext('webgl');
-      if (!ctx) return {
-        width: canvasSelector.width,
-        height: canvasSelector.height,
-        error: 'No WebGL context available' };
-
-      return {
-        width: canvasSelector.width,
-        height: canvasSelector.height,
-        contextAttributes: !!ctx.getContextAttributes(),
-        extensions: !!ctx.getSupportedExtensions()
-      };
-    });
-
-    // console.debug(canvasInfo);
-    return canvas;
-  };
 
   const verifyInitialState = async (page) => {
     const canvas = await waitForCanvas(page);
@@ -105,7 +33,6 @@ test.describe('Cord19 Canvas Tests', () => {
       test.slow();
       const canvas = await verifyInitialState(page);
 
-      // Perform zoom
       // Handle hover/tap based on device
       const isMobile = testInfo.project.name.includes('mobile');
       if (isMobile) {
